Supermarket_Data_Insight/deployment/requirements.txt
pandas
numpy
plotly.express
streamlit
scipy
matplotlib